requirements.txt
flask_restx
requests
urllib3
Pillow
black
slack_sdk
tableauserverclient
virtualenv

flask==1.1.2
itsdangerous==2.0.1
werkzeug==0.16.0
markupsafe==2.0.1
py3-validate-email==1.0.5
python-dateutil==2.8.1
Jinja2==2.11.3

seunggabi_core_python
